+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 2.12.0 (In Progress)
 ====================
 
+New Features
+------------
+
+### CallMessageEvents
+
+Users can now "opt-in" to their desired call message event types. "user-defined-message" will always be opted in by default.  
+- See [Device.Options.callMessageEvents](https://twilio.github.io/twilio-voice.js/interfaces/voice.device.options.html#callMessageEvents)
+
+#### Example
+
+```js
+// Create a Device instance that handles mutliple call message event types
+const options = {
+  callMessageEvents: ['user-defined-message', 'call-progress-event'],
+}
+const device = new Device(token, options);
+
+call.on('messageReceived', (message) => {
+  if (message.messageType === 'call-progress-event') {
+    console.log(JSON.stringify(message.content));
+    //the voiceEventSid can be used for tracking the message
+    console.log('voiceEventSid: ', message.voiceEventSid);
+  } else if (message.messageType === "user-defined-message") {
+    console.log(JSON.stringify(message.content));
+    //the voiceEventSid can be used for tracking the message
+    console.log('voiceEventSid: ', message.voiceEventSid);
+  }
+})
+```
+
 Changes
 -------